Rename seabird_extra_plugins package to extra
