+++ plugins/extra/mentions.go
@@ -1,4 +1,4 @@
-package seabird_extra_plugins
+package extra
 
 import (
 	"github.com/belak/go-seabird"
